RAP, PR, PO ->float qty [MIGRATE FRESH]

+++ database/migrations/2018_10_01_082922_create_trx_rap_detail_table.php
@@ -18,7 +18,7 @@ class CreateTrxRapDetailTable extends Migration
             $table->unsignedInteger('rap_id');
             $table->unsignedInteger('material_id')->nullable();
             $table->unsignedInteger('service_id')->nullable();
-            $table->integer('quantity');
+            $table->float('quantity',15,2);
             $table->double('price');
             $table->timestamps();
